Organise tab handling into files
