+++ src/completion/mod.rs
@@ -1,0 +1,7 @@
+mod base;
+mod default;
+mod tab_handler;
+
+pub use base::{ComplationActionHandler, Completer, Span};
+pub use default::DefaultCompleter;
+pub use tab_handler::DefaultCompletionActionHandler;
