+++ src/lib.rs
@@ -211,8 +211,8 @@ pub use highlighter::{DefaultHighlighter, Highlighter};
 mod styled_text;
 pub use styled_text::StyledText;
 
-mod completer;
-pub use completer::{
+mod completion;
+pub use completion::{
     ComplationActionHandler, Completer, DefaultCompleter, DefaultCompletionActionHandler, Span,
 };
